fix: internal hardening from code review
- Add (size_t) cast in malloc for free_request_stack (adaptive_ring.c)
- Use consistent memory_order_acq_rel for debug tick_entered (adaptive_engine.c)
- Guard adaptive_record_submit against negative sqe_count
- Add _Static_assert that MAX_SAMPLE_WINDOW >= MIN_SAMPLE_WINDOW
- Fix buffer_pool_free size=0 path: don't decrement total_buffers without
  decrementing total_allocated (just free directly on this error path)
- Add iteration bounds to hash table probe loops in buf_size_map

+++ engine/src/adaptive_ring.c
@@ -166,7 +166,7 @@ int ring_init(ring_ctx_t *ctx, int queue_depth, int cpu_id, const ring_options_t
     }
     memset(ctx->requests, 0, (size_t)queue_depth * sizeof(aura_request_t));
 
-    ctx->free_request_stack = malloc(queue_depth * sizeof(int));
+    ctx->free_request_stack = malloc((size_t)queue_depth * sizeof(int));
     if (!ctx->free_request_stack) {
         goto cleanup_requests;
     }
